refactor: move time constants to utils/ (fix data→ui layer violation)
THIRTY_MINUTES_MS / ONE_HOUR_MS lived in ui/constants.ts but were consumed
only by the data layer (sessionLiveness + all 5 providers). Move the four
time constants to a neutral src/utils/timeConstants.ts and repoint the six
importers. None of the four were used by the UI layer. Pure mechanical move,
no behavior change; obsolete layer-violation header notes removed.

Closes #192
