+++ src/ui/constants.ts
@@ -1,9 +1,8 @@
 /**
- * Shared UI constants — time thresholds, panel widths, terminal
- * limits, box-drawing characters — plus the display-width helpers
- * (`getDisplayWidth`, `getInnerWidth`, `createTitleLine`,
- * `createBottomLine`) used by every panel for borders and
- * truncation.
+ * Shared UI constants — panel widths, terminal limits, box-drawing
+ * characters — plus the display-width helpers (`getDisplayWidth`,
+ * `getInnerWidth`, `createTitleLine`, `createBottomLine`) used by
+ * every panel for borders and truncation.
  *
  * Design decision:
  * - `getDisplayWidth` is memoized across calls. Without
@@ -16,21 +15,7 @@
  *   No staleness invalidation is needed since display width is a
  *   pure function; eviction only caps memory, and an evicted hot
  *   string recomputes on its next miss.
- *
- * Gotcha:
- * - `THIRTY_MINUTES_MS` and `ONE_HOUR_MS` are imported by
- *   `src/data/sessions.ts` and `src/data/sessionLiveness.ts` —
- *   a data → ui layer violation flagged in those files. The
- *   constants should move to `src/utils/timeConstants.ts` (or
- *   live on `types/index.ts`) on a future refactor; keeping them
- *   here for now to avoid mixing the move with unrelated work.
  */
-
-// Time constants (in milliseconds)
-export const THIRTY_SECONDS_MS = 30 * 1000;
-export const THIRTY_MINUTES_MS = 30 * 60 * 1000;
-export const ONE_HOUR_MS = 60 * 60 * 1000;
-export const FIVE_MINUTES_MS = 5 * 60 * 1000;
 
 // Default panel width (can be overridden via config)
 export const DEFAULT_PANEL_WIDTH = 70;
